chore: merge config types to a single file
