+++ packages/core/src/types/thirdParty.ts
@@ -140,7 +140,6 @@ export interface CSSLoaderOptions {
   /**
    * Allow to enable/disables handling the CSS functions url and image-set.
    * If set to false, css-loader will not parse any paths specified in url or image-set
-   *
    * @default true
    */
   url?:
@@ -150,7 +149,6 @@ export interface CSSLoaderOptions {
       };
   /**
    * Allows to enables/disables @import at-rules handling.
-   *
    * @default true
    */
   import?:
@@ -175,20 +173,17 @@ export interface CSSLoaderOptions {
   /**
    * Allows to enables/disables or setups number of loaders applied before CSS loader for @import at-rules,
    * CSS Modules and ICSS imports, i.e. @import/composes/@value value from './values.css'/etc.
-   *
    * @default 0
    */
   importLoaders?: number;
   /**
    * By default, css-loader generates JS modules that use the ES modules syntax.
    * There are some cases in which using ES modules is beneficial, like in the case of module concatenation and tree shaking.
-   *
    * @default true
    */
   esModule?: boolean;
   /**
    * Allows exporting styles as array with modules, string or constructable stylesheet (i.e. CSSStyleSheet)
-   *
    * @default 'array'
    */
   exportType?: 'array' | 'string' | 'css-style-sheet';
@@ -205,13 +200,11 @@ export interface StyleLoaderOptions {
   /**
    * By default, style-loader generates JS modules that use the ES modules syntax.
    * There are some cases in which using ES modules is beneficial, like in the case of module concatenation and tree shaking.
-   *
    * @default true
    */
   esModule?: boolean;
   /**
    * Allows to setup how styles will be injected into the DOM.
-   *
    * @default 'styleTag'
    */
   injectType?: StyleLoaderInjectType;
